feat(modal): rules and settings modal
