+++ client/app/rules/rules.module.js
@@ -3,16 +3,16 @@ const angular = require('angular')
 const rulesService = require('./rules.service')
 const conditionList = require('./condition-list.component')
 const rulesRoute = require('./rules.route')
-const rulesManageRoute = require('./rules-manage.route')
+const rulesComponent = require('./rules.component')
 const run = require('./rules.run')
 
 const rulesModule = angular.module('simulator.rules', [
   require('angular-ui-router')
 ])
-  .factory('rulesService', rulesService)
-  .component('conditionList', conditionList)
   .config(rulesRoute)
-  .config(rulesManageRoute)
   .run(run)
+  .factory('rulesService', rulesService)
+  .component('rulesComponent', rulesComponent)
+  .component('conditionList', conditionList)
 
 module.exports = rulesModule
